LOD Springer
Interlinking LOD Springer data with other established conference
datasets

In the web of data the same real world entity can be described in
different datasets using different IDs and vocabularies. Interlinking
these descriptions is one of the main challenges of Linked Data, and is
crucial for data quality, discoverability and visibility.

Data described in Springer LOD Portal are not an exception, and it is
hardly possible to realise its full potential without interlinking it
with the rest of the LOD cloud. Improving the data quality via
interlinking is highly beneficial for such Linked Data applications as
data search and querying, visualization, analytics, data mining, content
enrichment, etc.

Technically, in the context of Linked Data, such interlinking is
realized using owl:sameAs links between URIs, which indicate that these
two URIs refer to the same thing. Creating and maintaining sameAs links
of high quality is a crucial aspect of Linked Data.

The goal of this project is to enrich the Springer Conference LOD data
with sameAs links to the existing resources, such as the ones listed
below. Additionally, the plan is to enrich the Springer data with
various conference ranking information.

Based on the data availability, potential impact and expected effort, we
propose to focus on the following data sources:
• ERA Australian Conference ranking and Brazilian Conference ranking by
QUALIS - sameAs links to series when possible + rank and its year
• DNB - sameAs links to conferences (conference example)
• Colinda conference portal - sameAs links to conferences
• Bonus task: DBpedia links for conference locations